klpsrc: add wrapper of patched functions
Patched functions might call each other. Without using
clang rewriter, we cannot blindly rename the patched
funcs. Instead, add wrappers to each of them.

There is a suppect here, when patched function is external,
we don't rename it so must still keep it external. Does this
conflict with the original global symbol in kernel? It seems
that the linker would auto translate the preserved global
to local symbol type which type is 't', the original is 'T',
and they all have the same name.