klpsrc: treat header or function scope static as new
A static variable defined in header or function is used
just in one function scope(mainly). It acts as some
state of the function.

For simplicity, now all treat this kind of static as
new variable. To some degree for patched function, its
state do in new state, seems that this choice is reasonable
for stateful function. But it is not clear whether in some
scenerio this would do harm to the kernel.